requirements.txt
PyPDF2==1.26.0
requests==2.23.0
Flask==1.1.2
Flask-Cors==3.0.8
stomp.py==7.0.0
PyMuPDF==1.17.7
opencv-python==4.1.0.25
imutils==0.5.2
numpy==1.17.0
py-zipkin==0.20.2
boto3==1.9.160
botocore==1.12.160
SQLAlchemy==1.3.6
pycryptodome==3.8.2
py-eureka-client==0.9.9
APScheduler==3.7.0
gunicorn==20.1.0
xmltodict==0.12.0
cx_oracle==8.1.0

